Add capability to drop literal from rule - equivalent to setting literal false.
This turns up, eg, in unit-literal propagation.  The clauses containing the
opposite-sense literal have said literal set to false - ie A OR B OR C
with ~A unit becomes FALSE OR B OR C - which simplifies to B OR C.

Dropping the literal achieves that effect while also speeding up operations
that are O (some function of N), so that's what I've done.