Add reverse-subsumption checks.
"Forward" subsumption is checking that candidate rule A is subsumed by
at least one existing rule - if so, it can be dropped without adding.

"Reverse" subsumption is checking if candidate rule A subsumes one or
more existing rules - if so, A still needs to be added, but the rules
A subsumes can be dropped.